rename snells examples
