+++ examples/snells_law.py
@@ -3,10 +3,9 @@
 import math
 from enum import Enum
 
-from magicgui import event_loop, magicgui
+from magicgui import magicgui
 
 
-# dropdown boxes are best made by creating an enum
 class Medium(Enum):
     """Enum for various media and their refractive indices."""
 
@@ -16,8 +15,8 @@ class Medium(Enum):
     Air = 1.0003
 
 
-@magicgui(call_button="calculate", result={"disabled": True, "fixedWidth": 100})
-def snells_law(aoi=30.0, n1=Medium.Glass, n2=Medium.Water, degrees=True):
+@magicgui(call_button="calculate", result_widget=True)
+def snells_law(aoi=1.0, n1=Medium.Glass, n2=Medium.Water, degrees=True):
     """Calculate the angle of refraction given two media and an AOI."""
     if degrees:
         aoi = math.radians(aoi)
@@ -30,12 +29,4 @@ def snells_law(aoi=30.0, n1=Medium.Glass, n2=Medium.Water, degrees=True):
         return "TIR!"
 
 
-with event_loop():
-    # the original function still works
-    # we can create a gui
-    gui = snells_law.Gui(show=True)
-    # we can list for changes to parameters in the gui
-    # ... these also trigger for direct parameter access on the gui object
-    gui.n1_changed.connect(print)
-    # we can connect a callback function to the __call__ event on the function
-    gui.called.connect(lambda x: gui.set_widget("result", str(x), position=-1))
+snells_law.show(run=True)
